+++ explorer/site-config/anvil/dev/config.ts
@@ -14,7 +14,6 @@ import { donorsEntityConfig } from "./Index/donorsEntityConfig";
 import { datasetsEntity } from "./datasetsEntity";
 import { filesEntity } from "./filesEntity";
 import { librariesEntityConfig } from "./Index/librariesEntityConfig";
-import { workspaceEntity } from "./workspaceEntity";
 
 // Images
 import logoAnvil from "images/logoAnvil.png";
@@ -71,7 +70,6 @@ const config: SiteConfig = {
     biosamplesEntity,
     librariesEntityConfig,
     activitiesEntity,
-    workspaceEntity,
     filesEntity,
   ],
   entityTitle: "Anvil Data Explorer",
