--- explorer/site-config/anvil/tsv-config.ts
@@ -1,63 +1,0 @@
-export const SOURCE_FIELD_KEY = {
-  BUCKET_NAME: "bucketName",
-  BUCKET_SIZE: "bucketSize",
-  COL: "COL",
-  CONSENT_LONG_NAME: "consentLongName",
-  CONSENT_TITLE: "consentTitle",
-  CONSORTIUM: "consortium",
-  DATATYPE: "library:datatype",
-  DATA_USE_RESTRICTION: "library:dataUseRestriction",
-  DISCOVERY_COUNT: "discoveryCount",
-  DISEASE_TEXT: "diseaseText",
-  DS: "DS",
-  FAMILY_COUNT: "familyCount",
-  GRU: "GRU",
-  GSO: "GSO",
-  HMB: "HMB",
-  INDICATION: "library:indication",
-  IRB: "IRB",
-  MDS: "MDS",
-  NAME: "name",
-  NPU: "NPU",
-  NRES: "NRES",
-  PARTICIPANT_COUNT: "participantCount",
-  PHSID: "phsId",
-  PUB: "PUB",
-  REQUESTOR_PAYS: "requestorPays",
-  SAMPLE_COUNT: "sampleCount",
-  STATUS: "status",
-  STUDY_DESIGN: "library:studyDesign",
-  SUBJECT_COUNT: "subjectCount",
-};
-
-export const SOURCE_FIELD_TYPE = {
-  BUCKET_NAME: "string",
-  BUCKET_SIZE: "number",
-  COL: "string",
-  CONSENT_LONG_NAME: "string",
-  CONSENT_TITLE: "string",
-  CONSORTIUM: "string",
-  DATATYPE: "string",
-  DATA_USE_RESTRICTION: "string",
-  DISCOVERY_COUNT: "number",
-  DISEASE_TEXT: "string",
-  DS: "string",
-  FAMILY_COUNT: "number",
-  GRU: "booleastring",
-  GSO: "string",
-  HMB: "string",
-  INDICATION: "string",
-  IRB: "string",
-  MDS: "string",
-  NAME: "string",
-  NPU: "string",
-  NRES: "string",
-  PARTICIPANT_COUNT: "number",
-  PHSID: "string",
-  PUB: "string",
-  REQUESTOR_PAYS: "boolean",
-  SAMPLE_COUNT: "number",
-  STATUS: "string",
-  STUDY_DESIGN: "string",
-  SUBJECT_COUNT: "number",
-};
